Use one-based column numbers.
Motivation:

Every programming editor I can find -- except emacs -- uses one-based column numbers. So the question is, do you think the majority of Javascript programmers in this day and age are using emacs or some other editor? If the majority of users are using a one-based editor, then it is much friendlier when you are reading an error report to have one-based column numbers.