Fix browserify support for loose and walk
We need the `require` call to be ignored in the first pass, which is
why it was inserted as `_dereq_` in the first place.  This is needed in
case the user is loading acorn and acorn_loose/acorn_walk as separate
script tags.

However, if the file is loaded via browserify (i.e. a second browserify
pass) we need that require to be processed, which wasn't working before
because it had been renamed to `_dereq_`.

This is fixed by using a placeholder, which gets swapped out after the
first browserify pass has completed.