Allow let being an identifier in ES2015 mode
This is more or less a hack as we modify the token after it is tokenized.

This causes the output of `onToken` and `tokenizer` to be different: the
former outputs a `name` token with value `let` while the latter outputs
a `let` token.

Fixes part of #227. `var yield` still fails to parse.