Implement template literal revision
The [template literal revision proposal](https://github.com/tc39/proposal-template-literal-revision) reached stage 4. This commit adds support for it when `ecmaVersion` is at least 9.

To allow bad unicode escapes to be tokenized, this updates the template tokenizer to fall back to tokenizing an `invalidTemplate` token if it encounters an invalid escape, rather than immediately throwing an error.