Fix frequent deopt in parseStatement
When checking for labeled statement, we were testing expr.type first, which meant always accessing another object shape (depending on node type) whenever a statement started with an identifier.

By simply reordering two checks, we can make sure that expr.type will be accessed only on the same node (Identifier) or on error, because it will be reached only if we saw a colon (":") right after it.