Forbid await as identifier in async arrow parameters
Unlike regular async functions, checks for async arrow functions are
delayed until we can ascertain that this is an async function. Hence,
separate processing is necessary. At the same time, the existing check
in this.toAssignable is not sufficient to check for some forbidden
cases, such as `async (a = (await) => b) => c`.

Also change "Can not" to the more grammatical "Cannot."