Revamp checkLVal()
Currently, the same checkLVal() function is used for quite a few
different purposes. While this arrangement has worked, pretty well one
can even say, we start to see that the lack of granularity checkLVal()
provides make the implementation of certain early errors harder than it
should be. For instance, the parser for for-loops has an ad-hoc error
condition for whenever the LHS is an AssignmentPattern. This condition
would be useful in quite a few other contexts, such as in
AssignmentExpressions, but it would be better to look for a more general
solution.

In this commit, we divide checkLVal() into three different functions for
different usage. This allows us to accurately account for the fact that
some lvalue types are allowed in certain contexts, but not others.
Additionally, add extensive documentation on how this approach is done.

The large-scale refactoring does mean that this PR would break
compatibility with existing plugins, and necessitate a major version
bump.

Fixes: #886