Update authors list, use .mailmap file
Use a git standard .mailmap file for removing duplicates and choosing
which names to use for authors that used different display names on
different commits.

This means `git log` and `git show` naturally pick it up, as well
as summary commands like `git shortlog -s` and `git shortlog -sn`,
and various GUIs for Git.

I've also added -f to the sort command as it appeared to (mostly)
be sorted case-insensitively already. Without this option, running
the script resulted in all lowercase names being moved to the end
of the list (tried both macOS/BSD's default sort, and the one from
gnu-coreutils). I'm guessing the previous run was done on a machine
with a different implementation or override present that uses some
kind of hybrid by default.