Add 2022 (and 13) as allowed ecmaVersion
Some features are only supported in 2022 and above. While some
of the existing tests already specify 2022 as ecmaVersion (or
the equivalent 13 level), it wasn't explicitly listed as allowed
option in the documentation or types.

Therefore, add the supported ecmaVersion to the list and types.
Downstream, we are using acorn and want to specify ecmaVersion
2022 for class field support, but TypeScript prohibits us from
doing that at the moment.