fix acorn-walk docs typo
In the acorn-walk readme, we say that `findNodeAt`'s `start` and `end` can be `null`, but our TS type definitions say `undefined` instead. This makes the readme match the TS types.

Interestingly, we actually _do_ explicitly check for `null` in the code, but we're doing double-equals rather than strict equality, so `undefined == null`.